size_t instead of int

+++ trunk/fflas-ffpack/ffpack/ffpack.h
@@ -174,7 +174,7 @@ namespace FFPACK { /* Permutations */
 	applyP( const Field& F,
 		const FFLAS::FFLAS_SIDE Side,
 		const FFLAS::FFLAS_TRANSPOSE Trans,
-		const size_t M, const int ibeg, const int iend,
+		const size_t M, const size_t ibeg, const size_t iend,
 		typename Field::Element_ptr A, const size_t lda, const size_t * P );
